% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_auc}
\alias{import_auc}
\title{Import pyscenic auc into R}
\usage{
import_auc(auc_scenic)
}
\arguments{
\item{auc_scenic}{A loom file return by pyscenic aucell.}
}
\value{
A list of following elements:
\itemize{
\item auc: AUC value matrix.
\item regulons: A list of regulon signatures attached with attributes binary AUC
value matrix ("binary_incidence") and thretholds ("threshold").
}
}
\description{
Import pyscenic auc into R
}
